refactor(lookup): Turn combined release lookup function into a class
The old function was too large to add a clean escape hatch for an
invalid GTIN, which is made easier by splitting it into multiple methods.

Since constructors can not be async, the extraction of a GTIN from the
first available result had to be dropped, but that was more of a hack
than a proper solution as it only works if the fastest provider delivers
a GTIN. The new implementation is more flexible and allows setting the
GTIN at any time to initiate more lookups.