feat(web): Add web application manifest which defines icons and colors
These are used when adding a website to the home screen or installing it
as a Progressive Web App (on Android).

Generate PNG icons in 192x192 and 512x512, these seem to be the relevant
sizes nowadays (Chrome, Firefox), SVG is only picked by Chrome so far.

Background is using the dark brand color, theme the red brand color.