style(web): Change manifest background color to white
Chrome always uses white as background for home screen icons and ignores
this attribute, so it makes sense to set it to white to make Firefox
home screen icons consistent.

I am intentionally not using pure white (#ffffff) as that apparently
causes Firefox to pick a red color form the icon!?

Also changing the them color (unused so far?) to the dark instead of the
red one to mitigate similar issues which might occur.