fix(merge): Preserve original provider errors and their stack
Previously the original errors had almost the same stack trace as the
`AggregateError` which was not very helpful for debugging.

An alternative would have been to set the cause of the `ProviderError`,
but this leads to the same (helpful) stack trace being logged twice.

So rather than wrapping every possible unintended exception with a
`ProviderError` we should pass it along and try to catch "intended"
exceptions (which should be wrapped) closer to their root cause.