fix(provider): Only update cache timestamp if it is newer
If the timestamp is not for the latest request, opening a permalink
causes all requests which happened after the timestamp not to be loaded
from a snapshot, but pointlessly being repeated instead.

Example failure scenario:

1) Lookup a Beatport release by URL.
2) Lookup the same Beatport release by barcode shortly afterwards.

The second lookup caches the search page, then uses the existing
snapshot for the release page from the first lookup and overwrites* the
newer search snapshot timestamp with the older release timestamp.

3) Opening the permalink, which contains the release timestamp, loads
only the release snapshot and requests the search results again.

*) Additionally the Beatport provider did not even set the cache
timestamp for the search request at all.