fix(Spotify): Update cache time after every search request
Precautionary fix, this is currently not an issue because web permalinks
will always contain a Spotify ID (since each GTIN search triggers a
consecutive ID lookup), so the cached search requests are not needed.

It is however possible to trigger a caching issue by doing a Spotify URL
lookup followed by a GTIN lookup of the same release.
Now the timestamp of the search request snapshot is higher than the one
of the other requests which will be used by the permalink.
If such a permalink is altered to contain a GTIN instead of the Spotify
ID, there is no matching snapshot for the search request, which causes
it to be repeated while all other cached requests are properly loaded.