feat(web): Drop `target='_blank'` from all external links
There are few good reasons (like interrupted playback, or loss of state)
to enforce links to open in a new tab, none of them apply to Harmony.
A user can still decide to open links in new tabs, forcing them to open
in the same tab is not easy when  `target='_blank'` is present.

Previously Harmony's external links were not even consistently using it.