refactor: Make name of resolvable entities optional
If an (optional) MBID is specified, we do not need the name, neither for
the existing usage for artist credits and labels, nor for future use
cases like recording and release group MBIDs.

Note: In order to display a label, we would still need the name, but
there can also be catalog numbers without an associated label.