fix(web): Prefer longest zero-padded GTIN variant for permalink
In an ideal world, a GTIN is just a number, but we have providers where
the correct zero-padding matters for lookups. Most of them already
handle this by doing multiple GTIN lookups if necessary, but Beatport
does not (because it would be way too slow with the HTML scraper).

By choosing the variant with the most zeros, more GTIN lookups should
succeed on first try, or work at all (for Beatport, when the "wrong"
GTIN variant without zero-padding has been encoded into the permalink).