feat(web): Allow to select providers for release actions
This is useful to avoid compatibility errors when the loaded release has
external links to incompatible releases.

For now this is not exposed to the UI because having provider checkboxes
here is just confusing for the regular use case after a redirect from MB
(where the URL contains the used/valid external IDs only).

Before we can use provider checkboxes here, we have to remove the flaw
that they lose ID values on form submission as they are boolean only.
(This is also the case for release permalinks.)