refactor(web): Decouple form value and persisted value of checkboxes
Since the corresponding query parameters for the provider checkboxes can
also be used to specify provider IDs (for permalinks), their form value
(which is sent for an enabled checkbox) can not be '1'.
Any non-empty value would be treated as a provider ID.

As a consequence, the form value for a provider checkbox has to be an
empty string. Until now the same value has also been used to persist the
checkbox state, but having the values '0' for false and '' for true was
not very intuitive, so I am changing the representation of true to '1'.

This is a non-breaking change for local storage values (which are
compared against '0'), but would have been a breaking change for cookie
values (which are compared against '' -- last chance to change that).