fix(test): Prevent fetch stubs from being restored too early
Currently they are restored as soon as the scope of the `describe`
function has been left, i.e. the tests were registered but not run yet.

This only worked in most cases because the stubbed `globalThis.fetch`
gets assigned to `MetadataProvider.fetch` which is mostly used, but
`globalThis.fetch` is no longer stubbed when the tests actually run!

Instead of letting the stubs be restored automatically with `using`, we
could do it manually, but that is not really necessary for any test.