feat(provider): Handle numeric HTTP Retry-After header for all responses
Might be necessary to avoid getting blocked by Spotify, but the logged
header may be interesting for other providers as well.

So far I have not encountered such a header for a successful request to
the Spotify API, only when it is too late already (HTTP 429 with a huge
Retry-After delay of multiple hours).