chore: Disable logging of HTTP error response bodies
These were originally added to debug iTunes HTTP 403 errors in
d3548cc32485303d516fb4b183e382c52f83b681, but their body is always empty.
Occasionally temporary issues of Beatport and Deezer are logged by this,
so we can reduce the verbosity until there is an actual need for it.