feat(web): Open all "Link external IDs" links
Adds a button to open all links to add external IDs for artists, labels,
and recordings. Since opening multiple tabs/windows is blocked by
default in most browsers, a warning is shown if the tabs were blocked
informing the user that they need to allow pop-ups for the site.

The button to open all links is only shown when more than one link is
available. Some margin around the group of links is added when there are
multiple links and the button is shown.

Other changes:
- 300ms delay after each opened tab to avoid ws/js rate limit errors
- New semantic class 'action' which uses the same styles as a 'message'
- Test cases for the extracted `getEditUrlToSeedExternalLinks` function